Some basic paging support.
Left/right/up/down for single line/char movements.
Page up and page down, or alternatively Ctrl-Up arrow and Ctrl-Down
arrow, for jumping up and down a screenfull at a time.
Ctrl-left and ctrl-right for moving left/right a screenfull at a time.
Home and end for going to top and bottom of output.